requirements.txt
accelerate==1.6.0
agent-starter-pack==0.3.8
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
arrow==1.3.0
asgiref==3.5.0
asn1crypto==1.5.1
async-timeout==4.0.3
asyncpg==0.30.0
attrs==23.2.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
bitsandbytes==0.45.5
blinker==1.4
build==1.2.2.post1
cachetools==5.5.2
certifi==2023.11.17
cffi==1.17.1
chardet==5.2.0
chromadb==1.0.7
cliapp==1.0.9
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.4
coloredlogs==15.0.1
contourpy==1.2.0
cookiecutter==2.5.0
cryptography==44.0.2
cycler==0.12.1
databases==0.9.0
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.7.0
durationpy==0.9
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.9
filelock==3.18.0
fiona==1.9.6
flatbuffers==25.2.10
fonttools==4.47.2
frozendict==2.4.0
frozenlist==1.6.0
fsspec==2025.3.2
gcsfs==2025.3.2
geopandas==0.14.3
google-api-core==2.24.2
google-auth==2.39.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.87.0
google-cloud-bigquery==3.31.0
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
greenlet==3.2.1
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.6
jeepney==0.7.1
Jinja2==3.1.6
joblib==1.3.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
keyring==23.5.0
kiwisolver==1.4.5
kubernetes==32.0.1
langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.56
langchain-text-splitters==0.3.8
langsmith==0.3.37
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.1.0
Markdown==3.3.6
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.8.2
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.4.3
networkx==3.4.2
nltk==3.9.1
numpy==1.26.3
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-nccl-cu12==2.26.2
oauthlib==3.2.2
onnxruntime==1.21.1
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
orjson==3.10.16
overrides==7.7.0
packaging==24.2
pandas==2.1.3
passlib==1.7.4
pdfminer.six==20250327
pdfplumber==0.11.6
peewee==3.17.0
pg8000==1.31.2
pgvector==0.4.1
pillow==10.2.0
posthog==4.0.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
Pygments==2.19.1
PyJWT==2.3.0
PyMuPDF==1.25.5
pyparsing==2.4.7
pypdfium2==4.30.1
PyPika==0.48.9
pyproj==3.6.1
PyQt5==5.15.10
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2023.4
PyYAML==6.0.2
regex==2024.11.6
requests==2.28.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.4.0
scipy==1.12.0
scramp==1.4.5
SecretStorage==3.3.1
sentence-transformers==4.1.0
sentencepiece==0.2.0
shapely==2.0.3
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.40
starlette==0.45.3
sympy==1.14.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.2.0
tokenizers==0.21.1
tomli==2.2.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
triton==3.3.0
tzdata==2023.4
unicorn==2.1.3
urllib3==1.24.2
uvicorn==0.34.2
uvloop==0.21.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
